requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1.5
django-python3-ldap==0.11.3
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
ldap3==2.8.1
numpy==1.19.5
openpyxl==3.0.6
pandas==1.1.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
sentry-sdk==1.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
